Move from system interface to component
