+++ app/src/components/v-license-badge.vue
